Add new protocol with client version announcement
This patch adds a new client->nuauth protocol. This protocol v24 is quiet similar to v22_1 but
it improves the phase between SASL authentication and cruise mode.

During this phase, all user messages are sent in response to a server message. Furthermore, the
server decide when the use has to switch to cruise mode (or warn user about failure).