Introduce support for plugins in libnuclient
Plugins (or modules) for libnuclient are shared libraries which can
extend libnuclient. Each plugin is instantiated with a unique name
(and thus the same shared libraries can be instantiated several times
with different parameters).
This implementation uses events (with a unique ID) and a dispatch function
in each plugin, to reduce the number of functions to be implemented.