Add new application: EMC
This adds a new application to NuFW suite, EMC. This application will be
used in multi-site context, acting as a directory to register nuauth server.
EMC is based on nussl/nuconfparser, and uses libev to handle connections
in an event loop.
Current implementation is in preliminary state: there are no threads,
incoming data are not processed, and there are a few problems to solve
(for ex, nussl_session_accept is blocking, and so will block the event
loop indefinitely).