Apply patch from Tomas Mraz
There is small mistake in internal.c where the buffer length for the
base64 encoded output is 3 bytes shorter than the value passed to
sasl_encode64. Fortunately the encoded output will not be ever so long
to write over these 3 bytes so the issue is not a real problem. However
the code should be fixed.