nuauth: fix race condition
If nuauth is sending the event back to the loop before releasing the
mutex, it is possible that user_worker is awaken by an other thread
before the release of the lock occurs. The effect was that under a
severe load, the server would not reinject the socket in the loop.